+++ examples/general/src/commonTest/kotlin/com/example/testLibrary/MochaStyle.kt
@@ -6,8 +6,8 @@ import de.infix.testBalloon.framework.core.TestSuite
 import de.infix.testBalloon.framework.core.TestSuiteScope
 import de.infix.testBalloon.framework.core.testSuite
 import de.infix.testBalloon.framework.shared.TestElementName
-import de.infix.testBalloon.framework.shared.TestElementPropertyFqn
 import de.infix.testBalloon.framework.shared.TestRegistering
+import de.infix.testBalloon.framework.shared.TestSuitePropertyName
 
 fun TestSuiteScope.describe(name: String, testConfig: TestConfig = TestConfig, content: TestSuite.() -> Unit) =
     testSuite(name, testConfig = testConfig, content = content)
@@ -22,6 +22,7 @@ fun TestSuiteScope.it(
 fun describe(
     @TestElementName name: String? = null,
     testConfig: TestConfig = TestConfig,
-    @TestElementPropertyFqn propertyFqn: String = "",
+    @TestSuitePropertyName qualifiedPropertyName: String = "",
     content: TestSuite.() -> Unit
-): Lazy<TestSuite> = testSuite(name = name, testConfig = testConfig, propertyFqn = propertyFqn, content = content)
+): Lazy<TestSuite> =
+    testSuite(name = name, testConfig = testConfig, qualifiedPropertyName = qualifiedPropertyName, content = content)