+++ testBalloon-framework-core/src/androidHostTest/kotlin/de/infix/testBalloon/framework/core/internal/integration/JUnit4ReportingTests.kt
@@ -102,7 +102,7 @@ class JUnit4ReportingTests {
 
     private fun descriptionTest(reportingMode: ReportingMode, expectedOutput: String): TestResult = try {
         FrameworkTestUtilities.withTestFramework(TestSession(reportingMode = reportingMode)) {
-            val suite1 by testSuite(propertyFqn = "c.e.topSuite") {
+            val suite1 by testSuite(qualifiedPropertyName = "c.e.topSuite") {
                 test("t-test 1") {}
 
                 testSuite("middle suite") {
@@ -122,7 +122,7 @@ class JUnit4ReportingTests {
                 }
             }
 
-            val suite2 by testSuite("top 2 named", propertyFqn = "c.e.topSuite2") {
+            val suite2 by testSuite("top 2 named", qualifiedPropertyName = "c.e.topSuite2") {
                 test("t2n-test1") {}
             }