+++ testBalloon-framework-core/src/commonTest/kotlin/de/infix/testBalloon/framework/core/TestSuiteTests.kt
@@ -26,7 +26,7 @@ class TestSuiteTests {
         val testCount = 3
         val expectedTestElementPaths = FrameworkTestUtilities.ConcurrentList<String>()
 
-        val suite by testSuite(propertyFqn = "topSuite") {
+        val suite by testSuite(qualifiedPropertyName = "topSuite") {
             for (suiteNumber in 1..subSuiteCount) {
                 testSuite("subSuite$suiteNumber") {
                     for (testNumber in 1..testCount) {
@@ -55,7 +55,7 @@ class TestSuiteTests {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
         val suite1 by testSuite(
-            propertyFqn = "suite1",
+            qualifiedPropertyName = "suite1",
             testConfig = TestConfig.aroundAll { tests ->
                 trace.add("$testElementPath aroundAll begin")
                 tests()
@@ -88,7 +88,7 @@ class TestSuiteTests {
 
     @Test
     fun aroundAllWithoutInnerInvocation() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1", testConfig = TestConfig.aroundAll {}) {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1", testConfig = TestConfig.aroundAll {}) {
             test("test1") {}
         }
 
@@ -104,7 +104,7 @@ class TestSuiteTests {
     @Test
     fun aroundAllWithoutInnerInvocationPermitted() = FrameworkTestUtilities.withTestFramework {
         val suite1 by testSuite(
-            propertyFqn = "suite1",
+            qualifiedPropertyName = "suite1",
             testConfig = TestConfig.permits(TestConfig.Permit.WrapperWithoutInnerInvocation).aroundAll {}
         ) {
             test("test1") {}
@@ -117,7 +117,7 @@ class TestSuiteTests {
 
     @Test
     fun lowerLevelSuiteOutOfHierarchy() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             val suite1 = this
             testSuite("childSuite") {
                 suite1.testSuite("child of wrong parent") {
@@ -134,8 +134,8 @@ class TestSuiteTests {
 
     @Test
     fun topLevelSuiteOutOfHierarchy1() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1") {
-            testSuite("top-level suite inside a top-level suite", propertyFqn = "inside suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
+            testSuite("top-level suite inside a top-level suite", qualifiedPropertyName = "inside suite1") {
             }
         }
 
@@ -149,9 +149,9 @@ class TestSuiteTests {
     @Test
     fun topLevelSuiteOutOfHierarchy2() = FrameworkTestUtilities.withTestFramework {
         fun myTopLevelSuite(name: String, content: TestSuite.() -> Unit) =
-            testSuite(name, propertyFqn = name, content = content)
+            testSuite(name, qualifiedPropertyName = name, content = content)
 
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             myTopLevelSuite("top-level suite inside a top-level suite") {
             }
         }
@@ -165,7 +165,7 @@ class TestSuiteTests {
 
     @Test
     fun suiteWithoutChildren() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
         }
 
         assertFails {
@@ -178,7 +178,7 @@ class TestSuiteTests {
     @Test
     fun suiteWithoutChildrenPermitted() = FrameworkTestUtilities.withTestFramework {
         val suite1 by testSuite(
-            propertyFqn = "suite1",
+            qualifiedPropertyName = "suite1",
             testConfig = TestConfig.permits(TestConfig.Permit.SuiteWithoutChildren)
         ) {
         }
@@ -189,19 +189,19 @@ class TestSuiteTests {
     }
 
     @Test
-    fun suiteWithEmptyPropertyFqn() = FrameworkTestUtilities.withTestFramework {
+    fun suiteWithEmptyQualifiedPropertyName() = FrameworkTestUtilities.withTestFramework {
         assertFails {
-            val suite1 by testSuite(propertyFqn = "") {
+            val suite1 by testSuite(qualifiedPropertyName = "") {
             }
             suite1
         }.also {
-            assertContains(it.message ?: "", "was not initialized with a non-empty 'propertyFqn'")
+            assertContains(it.message ?: "", "was not initialized with a non-empty 'qualifiedPropertyName'")
         }
     }
 
     @Test
     fun suiteWithBlankName() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = " ") {
+        val suite1 by testSuite(qualifiedPropertyName = " ") {
         }
 
         assertFails {
@@ -213,7 +213,7 @@ class TestSuiteTests {
 
     @Test
     fun testWithEmptyName() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             test("") {
             }
         }
@@ -227,7 +227,7 @@ class TestSuiteTests {
 
     @Test
     fun testWithBlankName() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             test(" ") {
             }
         }
@@ -244,7 +244,7 @@ class TestSuiteTests {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
         val suite1 by testSuite(
-            propertyFqn = "suite1",
+            qualifiedPropertyName = "suite1",
             testConfig = TestConfig
                 .aroundEach { elementAction ->
                     trace.add("$testElementPath aroundEach1.1 begin")
@@ -315,7 +315,7 @@ class TestSuiteTests {
 
     @Test
     fun failFast() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1", testConfig = TestConfig.failFast(3)) {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1", testConfig = TestConfig.failFast(3)) {
             for (testId in 1..15) {
                 test("test$testId") {
                     if (testId.mod(2) == 0) {
@@ -341,7 +341,7 @@ class TestSuiteTests {
     fun suiteLevelFixture() = FrameworkTestUtilities.withTestFramework {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
-        val invocationSuite by testSuite(propertyFqn = "invocationSuite") {
+        val invocationSuite by testSuite(qualifiedPropertyName = "invocationSuite") {
             val outerFixture =
                 testFixture { trace.also { it.add("$testElementPath fixture creating") } } closeWith
                     { trace.add("$testElementPath fixture closing") }
@@ -358,7 +358,7 @@ class TestSuiteTests {
             }
         }
 
-        val parameterSuite by testSuite(propertyFqn = "parameterSuite") {
+        val parameterSuite by testSuite(qualifiedPropertyName = "parameterSuite") {
             testFixture {
                 trace.also { it.add("$testElementPath fixture creating") }
             } closeWith {
@@ -377,7 +377,7 @@ class TestSuiteTests {
             }
         }
 
-        val contextSuite by testSuite(propertyFqn = "contextSuite") {
+        val contextSuite by testSuite(qualifiedPropertyName = "contextSuite") {
             testFixture {
                 trace.also { it.add("$testElementPath fixture creating") }
             } closeWith {
@@ -421,7 +421,7 @@ class TestSuiteTests {
     fun testLevelFixture() = FrameworkTestUtilities.withTestFramework {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
-        val parameterSuite by testSuite(propertyFqn = "parameterSuite") {
+        val parameterSuite by testSuite(qualifiedPropertyName = "parameterSuite") {
             testFixture {
                 trace.also { it.add("$testElementPath fixture creating") }
             } closeWith {
@@ -440,7 +440,7 @@ class TestSuiteTests {
             }
         }
 
-        val contextSuite by testSuite(propertyFqn = "contextSuite") {
+        val contextSuite by testSuite(qualifiedPropertyName = "contextSuite") {
             testFixture {
                 trace.also { it.add("$testElementPath fixture creating") }
             } closeWith {
@@ -484,7 +484,7 @@ class TestSuiteTests {
     fun testLevelFixtureNested() = FrameworkTestUtilities.withTestFramework {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
-        val parameterSuite by testSuite(propertyFqn = "parameterSuite") {
+        val parameterSuite by testSuite(qualifiedPropertyName = "parameterSuite") {
             testFixture {
                 1.also { trace.add("$testElementPath fixture '$it' creating") }
             } closeWith {
@@ -510,7 +510,7 @@ class TestSuiteTests {
             }
         }
 
-        val contextSuite by testSuite(propertyFqn = "contextSuite") {
+        val contextSuite by testSuite(qualifiedPropertyName = "contextSuite") {
             testFixture {
                 1.also { trace.add("$testElementPath fixture '$it' creating") }
             } closeWith {
@@ -566,7 +566,7 @@ class TestSuiteTests {
 
     @Test
     fun testLevelToSuiteLevelFixture() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             val fixture1 = testFixture {
             } asParameterForEach {
                 test("test1") { _ ->
@@ -590,7 +590,7 @@ class TestSuiteTests {
 
     @Test
     fun suiteLevelToTestLevelFixture() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             testFixture {
             } asParameterForAll {
                 test("test1") { _ ->
@@ -616,7 +616,7 @@ class TestSuiteTests {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
         val suite1 by testSuite(
-            propertyFqn = "suite1",
+            qualifiedPropertyName = "suite1",
             testConfig = TestConfig.aroundAll { tests ->
                 trace.add("$testElementPath aroundAll begin")
                 tests()
@@ -657,7 +657,7 @@ class TestSuiteTests {
     fun suiteLevelFixtureWithDisabledElements() = FrameworkTestUtilities.withTestFramework {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             val suite1Fixture =
                 testFixture { trace.also { it.add("$testElementPath fixture creating") } } closeWith
                     { trace.add("$testElementPath fixture closing") }
@@ -667,7 +667,7 @@ class TestSuiteTests {
             }
         }
 
-        val suite2 by testSuite(propertyFqn = "suite2") {
+        val suite2 by testSuite(qualifiedPropertyName = "suite2") {
             val suite2Fixture =
                 testFixture { trace.also { it.add("$testElementPath fixture creating") } } closeWith
                     { trace.add("$testElementPath fixture closing") }
@@ -681,7 +681,7 @@ class TestSuiteTests {
             }
         }
 
-        val suite3 by testSuite(propertyFqn = "suite3") {
+        val suite3 by testSuite(qualifiedPropertyName = "suite3") {
             val suite3Fixture =
                 testFixture { trace.also { it.add("$testElementPath fixture creating") } } closeWith
                     { trace.add("$testElementPath fixture closing") }
@@ -719,7 +719,7 @@ class TestSuiteTests {
     fun suiteLevelFixtureWithFailedTest() = FrameworkTestUtilities.withTestFramework {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             val fixture1 =
                 testFixture { trace.also { it.add("$testElementPath fixture creating") } } closeWith
                     { trace.add("$testElementPath fixture closing") }
@@ -774,7 +774,7 @@ class TestSuiteTests {
         }
 
         fun testSuiteWithTestLevelFixture(name: String, testSuccesses: List<Boolean>): TestSuite =
-            testSuite(propertyFqn = "test-level, $name") {
+            testSuite(qualifiedPropertyName = "test-level, $name") {
                 testFixtureChecking() asParameterForEach {
                     testSuccesses.forEachIndexed { index, testSuccess ->
                         test("test-$index") {
@@ -787,13 +787,13 @@ class TestSuiteTests {
             }.value
 
         fun testSuiteWithSuiteLevelFixture(name: String, testSuccesses: List<Boolean>): TestSuite =
-            testSuite(propertyFqn = "suite-level, $name") {
+            testSuite(qualifiedPropertyName = "suite-level, $name") {
                 val fixture = testFixtureChecking()
                 testSeries(suiteLevelFixture = fixture, testSuccesses)
             }.value
 
         fun testSuiteWithSuiteLevelFixtureAndSubSuite(name: String, testSuccesses: List<Boolean>): TestSuite =
-            testSuite(propertyFqn = "suite-level+sub, $name") {
+            testSuite(qualifiedPropertyName = "suite-level+sub, $name") {
                 val fixture = testFixtureChecking()
                 testSuite("sub") {
                     testSeries(suiteLevelFixture = fixture, testSuccesses)
@@ -801,7 +801,7 @@ class TestSuiteTests {
             }.value
 
         fun testSuiteWithSuiteLevelFixtureAndSubSubSuite(name: String, testSuccesses: List<Boolean>): TestSuite =
-            testSuite(propertyFqn = "suite-level+sub+sub, $name") {
+            testSuite(qualifiedPropertyName = "suite-level+sub+sub, $name") {
                 val fixture = testFixtureChecking()
                 testSuite("sub") {
                     testSuite("sub") {
@@ -863,7 +863,7 @@ class TestSuiteTests {
         var failCount = 0
         var closeCount = 0
 
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             val fixture1 =
                 testFixture { fail("fixture failing intentionally (${++failCount})") } closeWith { closeCount++ }
 
@@ -893,7 +893,7 @@ class TestSuiteTests {
     fun suiteLevelFixtureWithSetupFailures() = FrameworkTestUtilities.withTestFramework {
         val trace = FrameworkTestUtilities.ConcurrentList<String>()
 
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             val fixtures = listOf(
                 testFixture { trace.add("$testElementPath fixture1 creating") } closeWith
                     { trace.add("$testElementPath fixture1 closing") },
@@ -1017,7 +1017,7 @@ class TestSuiteTests {
 
     @Test
     fun uniqueTestPaths() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(propertyFqn = "suite1") {
+        val suite1 by testSuite(qualifiedPropertyName = "suite1") {
             test("test1") {
             }
 
@@ -1038,12 +1038,12 @@ class TestSuiteTests {
 
     @Test
     fun uniqueTopLevelSuitePathsInSameCompartment() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(name = "common name", propertyFqn = "suite1") {
+        val suite1 by testSuite(name = "common name", qualifiedPropertyName = "suite1") {
             test("test1") {
             }
         }
 
-        val suite2 by testSuite(name = "common name", propertyFqn = "suite2") {
+        val suite2 by testSuite(name = "common name", qualifiedPropertyName = "suite2") {
             test("test1") {
             }
         }
@@ -1055,12 +1055,12 @@ class TestSuiteTests {
 
     @Test
     fun uniqueTopLevelSuitePathsInDifferentCompartments() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite(name = "common name", propertyFqn = "suite1") {
+        val suite1 by testSuite(name = "common name", qualifiedPropertyName = "suite1") {
             test("test1") {
             }
         }
 
-        val suite2 by testSuite(name = "common name", propertyFqn = "suite2", compartment = {
+        val suite2 by testSuite(name = "common name", qualifiedPropertyName = "suite2", compartment = {
             TestCompartment.Concurrent
         }) {
             test("test1") {
@@ -1097,7 +1097,10 @@ class TestSuiteTests {
         val additionalReportA = AdditionalExecutionReport("A")
         val additionalReportB = AdditionalExecutionReport("B")
 
-        val suite1 by testSuite(propertyFqn = "suite1", testConfig = TestConfig.executionReport(additionalReportA)) {
+        val suite1 by testSuite(
+            qualifiedPropertyName = "suite1",
+            testConfig = TestConfig.executionReport(additionalReportA)
+        ) {
             test("test1", testConfig = TestConfig.disable()) {
             }