+++ testBalloon-framework-core/src/jvmTest/kotlin/de/infix/testBalloon/framework/core/TestConfigTestsJvm.kt
@@ -21,7 +21,7 @@ class TestConfigTestsJvm {
             @OptIn(ExperimentalCoroutinesApi::class)
             withSingleThreadedDispatcher { defaultDispatcher ->
                 val testSuite by testSuite(
-                    propertyFqn = "testSuite",
+                    qualifiedPropertyName = "testSuite",
                     testConfig = TestConfig.coroutineContext(defaultDispatcher)
                 ) {
                 }
@@ -52,7 +52,7 @@ class TestConfigTestsJvm {
     private suspend fun testMainDispatcher() {
         initializeTestFramework()
         try {
-            val testSuite by testSuite(propertyFqn = "testSuite") {}
+            val testSuite by testSuite(qualifiedPropertyName = "testSuite") {}
 
             // Using the platform's main dispatcher.
             val platformMainThreadId = withContext(Dispatchers.Main) { testPlatform.threadId() }
@@ -100,7 +100,7 @@ class TestConfigTestsJvm {
     @Test
     fun mainDispatcherSetMoreThanOnce() {
         FrameworkTestUtilities.withTestFramework {
-            val testSuite by testSuite(propertyFqn = "testSuite") {}
+            val testSuite by testSuite(qualifiedPropertyName = "testSuite") {}
 
             @OptIn(ExperimentalCoroutinesApi::class)
             withSingleThreadedDispatcher { alternativeMainDispatcher ->